pending/vision_engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-engine"
version = "0.1.0"
description = "Universal vision AI engine using Ollama vision models"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "pillow>=10.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=23.0", "ruff>=0.1.0"]
